Update Version, code cleanup
