+++ HttpClientDecorator.Tests/HttpGetCallServiceTelemetryTests.cs
@@ -1,72 +1,6 @@
-﻿using HttpClientDecorator.Interfaces;
-using Microsoft.Extensions.Logging;
-using Moq;
+﻿namespace HttpClientDecorator.Tests;
 
-namespace HttpClientDecorator.Tests
+[TestClass]
+public class HttpGetCallServiceTelemetryTests
 {
-    [TestClass]
-    public class HttpGetCallServiceTelemetryTests
-    {
-        private Mock<IHttpClientService> _mockService;
-        private Mock<ILogger<HttpClientSendServiceTelemetry>> _mockLogger;
-        private HttpClientSendServiceTelemetry _telemetryService;
-
-        [TestInitialize]
-        public void TestSetup()
-        {
-            _mockService = new Mock<IHttpClientService>();
-            _mockLogger = new Mock<ILogger<HttpClientSendServiceTelemetry>>();
-            _telemetryService = new HttpClientSendServiceTelemetry(_mockLogger.Object, _mockService.Object);
-        }
-
-        [TestMethod]
-        public async Task GetAsync_ReturnsExpectedResult()
-        {
-            // Arrange
-            var expectedResponse = new HttpClientSendRequest<string>
-            {
-                RequestPath = "https://example.com",
-                ResponseResults = "OK",
-                Retries = 0
-            };
-            _mockService.Setup(x => x.HttpClientSendAsync(It.IsAny<HttpClientSendRequest<string>>(), It.IsAny<CancellationToken>()))
-                .ReturnsAsync(expectedResponse);
-
-            // Act
-            var result = await _telemetryService.HttpClientSendAsync(expectedResponse, CancellationToken.None);
-
-            // Assert
-            Assert.AreEqual(expectedResponse.RequestPath, result.RequestPath);
-            Assert.AreEqual(expectedResponse.ResponseResults, result.ResponseResults);
-            Assert.AreEqual(expectedResponse.Retries, result.Retries);
-            Assert.IsNotNull(result.CompletionDate);
-            Assert.IsTrue(result.ElapsedMilliseconds >= 0);
-            _mockLogger.Verify(x => x.Log(It.IsAny<LogLevel>(), It.IsAny<EventId>(), It.IsAny<It.IsAnyType>(), It.IsAny<Exception>(), (Func<It.IsAnyType, Exception, string>)It.IsAny<object>()), Times.Never);
-        }
-
-        [TestMethod]
-        public async Task GetAsync_LogsErrorAndReturnsOriginalResult_WhenExceptionIsThrown()
-        {
-            // Arrange
-            var expectedResponse = new HttpClientSendRequest<string>
-            {
-                RequestPath = "https://example.com",
-                ResponseResults = null,
-                Retries = 0
-            };
-            var expectedException = new Exception("Something went wrong");
-            _mockService.Setup(x => x.HttpClientSendAsync(It.IsAny<HttpClientSendRequest<string>>(), It.IsAny<CancellationToken>()))
-                .ThrowsAsync(expectedException);
-
-            // Act
-            var result = await _telemetryService.HttpClientSendAsync(expectedResponse, CancellationToken.None);
-
-            // Assert
-            Assert.AreEqual(expectedResponse.RequestPath, result.RequestPath);
-            Assert.IsNull(result.ResponseResults);
-            Assert.AreEqual(expectedResponse.Retries, result.Retries);
-            Assert.IsNotNull(result.CompletionDate);
-            Assert.IsTrue(result.ElapsedMilliseconds >= 0);
-        }
-    }
 }
