+++ HttpClientDecorator/HttpClientSendPollyOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace HttpClientDecorator;
+
+public class HttpClientSendPollyOptions
+{
+    public int MaxRetryAttempts { get; set; }
+    public TimeSpan RetryDelay { get; set; }
+    public int CircuitBreakerThreshold { get; set; }
+    public TimeSpan CircuitBreakerDuration { get; set; }
+}
